MHKiT-MATLAB Test Release

+++ mhkit/power/characteristics/instantaneous_frequency.m
@@ -36,7 +36,7 @@ if dsize(2)>1
    end
    data_pd=py.mhkit_python_utils.pandas_dataframe.spectra_to_pandas(time(:,1),li,int32(dsize(2)));
 elseif dsize(2)==1
-   data_pd=py.mhkit_python_utils.pandas_dataframe.spectra_to_pandas(time,py.numpy.array(data),int32(dsize(2)));
+   data_pd=py.mhkit_python_utils.pandas_dataframe.spectra_to_pandas(time,py.numpy.array(data),dsize(2));
 end
 
 harmonics_pd = py.mhkit.power.characteristics.instantaneous_frequency(data_pd);
